+++ packages/reactor-extension/src/view/configuration/brandConciergeSection.jsx
@@ -85,7 +85,6 @@ export const bridge = {
         instanceSettings.conversation = conversation;
       }
     }
-
     return instanceSettings;
   },
 
@@ -94,9 +93,10 @@ export const bridge = {
       is: true,
       then: (conciergeSchema) =>
         conciergeSchema.shape({
-          region: string().matches(
-            /^[a-z]{2,4}[0-9]{1,2}$/i,
+          region: string().test(
+            "valid-region",
             "Please enter a valid region (e.g. va7, or2, irl1).",
+            (value) => !value || /^[a-z]{2,4}[0-9]{1,2}$/i.test(value),
           ),
           stickyConversationSession: boolean(),
           streamTimeout: number()